Creating list finished (i hope!)

+++ app/src/main/java/pt/isec/amovtp1/grocerylistmanagement/data/Data.kt
@@ -3,12 +3,11 @@ package pt.isec.amovtp1.grocerylistmanagement.data
 import java.util.*
 import kotlin.collections.ArrayList
 
-// date = Calendar.getInstance().time
 class ListInformation(var listName: String, var date: Date)
 
 class Product(var name: String, var category: String, var brand: String?, var filePath: String, var observations: ArrayList<Observation>) {
     // Amount Class
-    data class Amount(var quantity: Int, var unit: String)
+    data class Amount(var quantity: Double, var unit: String)
     var amount: Amount? = null
 
     // Observations Class
@@ -18,7 +17,7 @@ class Product(var name: String, var category: String, var brand: String?, var fi
     data class Price(var price: Double, var priceDate: Date)
     var prices: ArrayList<Price> = arrayListOf()
 
-    constructor(name: String, category: String, brand: String?, filePath: String, observations: ArrayList<Observation>, quantity: Int, unit: String)
+    constructor(name: String, category: String, brand: String?, filePath: String, observations: ArrayList<Observation>, quantity: Double, unit: String)
             : this(name, category, brand, filePath, observations) {
         amount = Amount(quantity, unit)
     }
